Wrote failing specs which highlights missing content-type with rack test driver
Don't know if this could have been written any better, guess it can as
this is my first "deep" look into Capybara. The problem is found in the
Capybara::Driver::RackTest; it doesn't forward the content type of the
file being uplloaded to Rack::Test::UploadedFile.new. UploadFile's
initializer sets the content type to text/plain as a default.

This will make for instance Paperclip's attachment validation fail when
testing with Cucumber and Capybara even though we are uploading a type
which should pass.