This adds a server_host option to Capybara.
In order to run Selenium via Capybara on my virtual machine, you need to be
able to bind the server IP to a net interface on the same subnet as the VM.

This patch exposes an configuration variable to allow you to bind the
test server to any net interface on your local machine.

Falls back to 127.0.0.1 if the config var is not set.