Don't follow external redirects for the rack-test driver.
Previously the Capybara::RackTest::Driver would follow all redirects even those to external URLs. This would cause exceptions to be raised if the URL didn't match a route in the application being tested. This is because rack-test can't hit external URLs. The expected usage here is that the developer just wants to test that the redirect went to the correct external location not test anything about it. If they actually want to make assertion about the content at the external URL or interact with the external page in some way they should use an alternate driver.

This adds the `#redirect_url` method to the Capybara::Session object. This way the developer can test that the app redirected to the correct location:

    visit('/redirect_to_external_location')
    redirect_url.should == 'http://www.google.com'

This new behavior can be changed back to the old behavior by passing false to the `follow_external_redirects` option to the Capybara::RackTest::Driver:

    Capybara.register_driver :rack_test do |app|
      Capybara::RackTest::Driver.new(app, :follow_external_redirects => true)
    end