Expand Test::Unit README section
Call reset_sessions! and use_default_driver in sample code, and explain
driver switching.

For Rails, be opinionated and recommend using DatabaseCleaner with the
truncation strategy. (Selective switching to transaction cleaning for
Rack::Test is hard to do without metadata, and the shared connection
thing feels a bit hackish -- you never know when it might break in the
future.)