Remove broken tests for server_boot_timeout
When they were added in 09d6335, responsive? already only hit
/__identify__, which does not hit the server, but gets special-cased in
Capybara's Identify server class (and runs instantly).

So it's not clear what constitutes a "slow" server that might time out
at boot, and thus why server_boot_timeout needs to be configurable. A
server that is slow to respond to HTTP requests would not trigger the
boot timeout. All we really seem to be waiting for within the timeout
window is for Thin or WEBrick to come up.