Proof of concept for `find` failing on ambiguous matches
Just so we have some code to test with our real-life apps. :)

An "unambiguous" match is one with exactly one matching node or exactly
one matching visible node.

Error messages are not updated, it just says "not found" if there is
more than one match.

The test suite of course does not pass with this.