Use method="post".
WebKit will append a '?' to the URL when submitting a form via GET, even
if there is no actual form data. (Firefox does not.) This causes a couple
of tests to fail for the Poltergeist driver, because the tests do not
expect there to be a '?'. Presumably it would also affect
capybara-webkit.

I am not sure which behaviour is 'correct' but it would seem a bad idea
to mess with the URL that the browser is actually providing, so changing
to test like this just allows the results to be consistent.