Correct the specs for Session#body, #html, #source.
Per f4360f60998ea4805dfba89a1613939d054975c4, #body should return the
unmodified source, just like #source. However, in Selenium it doesn't,
which is why this test was passing, even though it is incorrect.

I have corrected the spec for #body to match #source (as it is just an
alias), and changed the original spec for #body to be the spec for #html.