For Capybara's own test suite, keep default_wait_time always at 1
Until now, we had it set at 0 for all tests that did't require `:js`.
But that would cause flickering failures when tests that didn't require
`:js` ran against Selenium, e.g. like this:

    visit '/foo' # asynchronous in Selenium
    page.should have_content 'Foo' # can fail if visit didn't finish in time

This change doubles(!) test suite run time, presumably because we are
testing for expected failures in a lot of places, causing us to retry
for the full default_wait_time period. At this point though I'm not sure
if there's a simple way to avoid the performance hit.