Improve failure message when text is present but invisible.
When testing a page with JS that hides and shows 
certain elements, it's useful to know whether a 
test failure means that an element has failed to 
render, or has rendered but become invisible (or 
failed to become visible). 

For instance, some jQuery plugins will take 
existing elements and hide them, adding some other
elements onto the page that use the originals as
storage. If your tests are written to use the 
now-hidden elements they will mysteriously start
failing. 

This commit only changes the behavior for failed
tests (rspec and assert) and only for text
queries. It might be nice to extend this to the 
other finders.