Temporarily use different options for a driver.
When testing an action that redirects offsite it is useful to be able to not follow redirects just temporarily when it is enough to assert that a redirection has occured to the correct location.

Currently when using the Capybara::RackTest::Driver an attempt to redirect to an external site results in a request with the same path to the app currently under test. Most of the time this will cause a test to fail as a request is made to a non-existent path.