[fix] selenium set_text passing in empty string with options
If an empty value if passed in with a clear option, it seems that the
caller's intent is for the option to be honored regardless of the value
to be set in the field.

Just passing in a blank value would behave as it did before.

Updated spec to also attach an input event listener to
`with_change_event`