Test and document has_select? for no selected options.
Typically, we use `has_select? 'foo', selected: 'a'` to determine if a select
element 'foo' has an option 'a' with the `selected` attribute. However, we can
also invoke `has_select? 'foo', selected: []` to determine that none of the
options has the `selected` attribute. This is handy, but not documented nor
tested.

Test and document `has_select? 'foo', selected: []`.