Retry navigation if unload alert disappears
When a Selenium driver interaction raises an unexpected alert error,
Firefox 59+ [automatically dismisses the user prompt][1] that caused it.
Therefore, if the driver reset triggers an unload alert, we cannot wait
to handle it until we receive an unexpected alert error. Instead, we
must retry navigation and proactively attempt to accept any alert that
appears.

[1]: https://bugzilla.mozilla.org/show_bug.cgi?id=1416284