Fix default puma server if puma gem is not present
Currently when upgrading to Capybara 3, the puma server is the default
server regardless if puma is present in the test application itself.

This causes the following error to be raised in non-puma apps test
suites:

```
LoadError:
  cannot load such file -- rack/handler/puma
/gems/activesupport-4.2.10/lib/active_support/dependencies.rb:274:in `require'
/gems/activesupport-4.2.10/lib/active_support/dependencies.rb:274:in `block in require'
/gems/activesupport-4.2.10/lib/active_support/dependencies.rb:240:in `load_dependency'
/gems/activesupport-4.2.10/lib/active_support/dependencies.rb:274:in `require'
/gems/capybara-3.0.0/lib/capybara.rb:441:in `block in <top (required)>'
/gems/capybara-3.0.0/lib/capybara.rb:235:in `run_default_server'
/gems/capybara-3.0.0/lib/capybara.rb:432:in `block in <top (required)>'
/gems/capybara-3.0.0/lib/capybara/server.rb:103:in `block in boot'
```

This requires the following modification to be made to the app to get it
to work again.

```
Capybara.server = :webrick
```

Instead, check if puma is present and only then default to puma as the
default server. This will make the upgrade path easier for those
upgrading to 3.0.0 and running alternative servers such as Unicorn.