Use `load_selenium` class method for requiring `selenium-webdriver`.
This change relates to this issue:

https://github.com/teamcapybara/capybara/issues/2011

`Capybara::Selenium::Driver#load_selenium` is bumped to a class method which is
called either during `Selenium::Driver` instantiation or before Selenium is
used. Tests have also been updated to remove now unnecessary requires of
`selenium-webdriver`.

This should make `Selenium::Driver` behave more like `RackTest::Driver`. It also
has the benefit of making sure the misnamed Selenium errors are patched
~everywhere.

This change passes local tests for me that were run with:

`bundle exec rake chrome_spec`

`CI=true bundle exec rake travis`

All in all, I'm not sure the lazy-loading is worthwhile (though that's, of
course, a choice for the maintainers).

A small aside, that there's also the use of the `webdrivers` gem which is
conditional on the `CI` environment variable. This means that CI behaves
differently to other environments in a way that's perhaps not expected (though
the failure case looks to be that tests would succeed in CI but fail locally,
which seems better than the opposite).

Maintainers, if this isn't the direction you want to go feel free to close the
pull request without merging.