Fix regression in 3.31.0 with Result#[] using inclusive range ending in -1
The logic was changed to support beginless ranges, and it broke
this case, because later in the method it loads up to max + 1,
which will be 0 for inclusive ranges ending in -1.  Since in
this context, inclusive ranges ending in -1 are endless, treat
them as endless ranges.

Not sure if this is the best way to fix this, but it does work.