Add support for custom message to `assert_selector`
As a developer writing tests, the built-in error messages
for assert_selector can sometimes be confusing/not clear.

Allow for custom error messages for `assert_selector`.

This PR is just an idea. I could see us adding this option
to all the assertion helpers in Capybara, but wanted to get
feedback on the idea first :heart:.