Find <button> elements through <label>
Browsers consider [<button> elements][mdn-button] as
["labelable"][labelable], meaning that a `<button
id="button-with-label">` element can be labelled by a [`<label
for="button-with-label">` element][mdn-label].

This commit extends the button selector to resolve `<button>` elements
with a locator that references a `<label>` element's text. That
`<label>` element can either reference the button through its `[for]`
attribute, or it can be an ancestor of the `<button>` element.

[labelable]: https://developer.mozilla.org/en-US/docs/Web/Guide/HTML/Content_categories#Form_labelable
[mdn-button]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/button
[mdn-label]: https://developer.mozilla.org/en-US/docs/Web/HTML/Element/label