Limit ARIA role check on buttons
Add test coverage to ensure that definitions and selectors for other
buttons aren't affected by the `enable_aria_role:` option.

The `:button` selector definition becomes too lax when
`Capybara.enable_aria_role = true`, which results in [false negatives
for `have_no_button`][false-negative] and false positives for
`have_button`.

The source is related to when that XPath is combined with other
selectors in the [Selector::Definition::Button][].

The lines were last changed as part of #2400, so this commit [partially
modifies some of those changes][comment].

[false-negative]: https://travis-ci.org/github/teamcapybara/capybara/jobs/746824288#L868-L871
[Selector::Definition::Button]: https://github.com/teamcapybara/capybara/blob/30ec8c188dec85f3e316f1f60d05574033b3186d/lib/capybara/selector/definition/button.rb#L22-L25
[comment]: https://github.com/teamcapybara/capybara/pull/2400#issuecomment-700864666?