Make animation disabler threadsafe
Animation Disabler is installed as a rack middleware - it will be
instantiated once and the call method will be called by multiple
requests.

The code was using instance variables on the middleware for information
about individual requests. When multiple requests were made at the same
time, the subsequent request would replace the status/headers/body of
the original request, causing the original request to return the
response of the subsequent request.