Use the new headless Chrome on the pre-registered driver (#2726)
Earlier this year (2023), Chrome released Chrome for Testing, a new
Chrome flavor that targets the testing and automation use cases. This
flavor is accessible using the `--headless=new` option since Chrome 109,
released January 1, 2023. This new mode supports the full browser
capabilities, including running extensions.

The default headless mode is still pointed to `--headless=old` for now,
but it's planned to be changed to `new` over time.

This commit explicitly chooses the `new` headless mode, that uses Chrome
for Testing.