Expand on `Capybara::Selector` documentation (#2736)
The [Capybara::Selector][] documentation highlights the wealth of
capabilities that Capybara provides out of the box with its built-in
selectors.

While they're powerful, they're often misunderstood. This adds a
dedicated heading to the README to elaborate more on how they work
across the entire system. The main aim is to highlight how most methods
support the trio of Selector arguments: name, locator, and filters.

[Capybara::Selector]: https://rubydoc.info/github/teamcapybara/capybara/master/Capybara/Selector

Co-authored-by: Matijs van Zuijlen <matijs@matijs.net>